feat: Hide 'Buy Car' button at Bank
The 'Buy Car' button was incorrectly being shown when the player traveled to the Bank. This commit fixes the bug by removing the code that was making the button visible.

The corresponding UI test has also been updated to assert that the 'Buy Car' button is hidden at the Bank.